Remove newState when not needed
Enhanced code readability. Any reasonable C++ optimizer would notice
newStats isn't changed and pass it as a constant. The problem is why to
bother. Having a variable gives developer hint this might change
somewhere in the code before being used. Usually this isn't the case and
its use seems as a kind of scaffolding for future changes.

It's much better to inform developer this is the end state so he/she
doesn't have to look for newState assignment.