fixed synchronization problem in ControllerThread
When a change is detected in the std::queue, and our intend is to
execute all the commmands in the queue, there must be a while loop
extracting queue memebers before queue is empty. Otherwise, if two (or
more) elements are queued in while queue lock is being hold, only the
first one will be retrieved and executed. The later ones will remain in
queue as there will be no routine to take them out.