Fixes various small issues with FPGA/NI RT code (#40)
* NiOpen macro

Create macro properly reporting which parameters were used when opening
a bitfile.

* newly generated FPGA bitfiles

* updated RIO reference

* adds -s option setting SAL debug level

* finalize in FPGA destructor

* serial status header

* fixed synchronization problem in ControllerThread

When a change is detected in the std::queue, and our intend is to
execute all the commands in the queue, there must be a while loop
extracting queue members before queue is empty. Otherwise, if two (or
more) elements are queued in while queue lock is being hold, only the
first one will be retrieved and executed. The later ones will remain in
queue as there will be no routine to take them out.

* prints warning when subscriber thread takes too long to execute

* verbose command

* Reset force actuator bump test status on startup

Tested actuator ID must be rest to 0 on startup.

* set/getPower CLI commands

* set bump test force tolerance to 5 N

* clang formatted, fixed includes

* fixed modbus CRC calculation