fixes timeout while mirror was lowered after aborting raise
The original implementation uses shared variable for both raise and
lowering timeouts. As class handling mirror raise and lowering was split
to two, the lowering timestamp wasn't properly updated. This resulted in
mirror being faulted right after it started to be lowered due to abort.
This commit fixes the problem by providing public function in lowering
controller to setup mirror for lowering following abort in raise.